Add checkbox for running test in GH build workflow (#4246)
Currently when we run a build with limited set of boards that doesn't include
OVA, the test job fails because the OVA artifact is missing. Add a checkbox for
running tests and ensure that OVA artifact is built if it's enabled.

(cherry picked from commit 71d222ad91f8c651c7a640ff8530fff7b7fec5d9)